+++ crates/biome_cli/src/execute/migrate/eslint_any_rule_to_biome.rs
@@ -885,7 +885,7 @@ pub(crate) fn migrate_eslint_any_rule(
             }
             let group = rules.nursery.get_or_insert_with(Default::default);
             let rule = group
-                .use_consistent_builtin_instatiation
+                .use_consistent_builtin_instantiation
                 .get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }
@@ -1285,7 +1285,7 @@ pub(crate) fn migrate_eslint_any_rule(
             }
             let group = rules.nursery.get_or_insert_with(Default::default);
             let rule = group
-                .use_consistent_builtin_instatiation
+                .use_consistent_builtin_instantiation
                 .get_or_insert(Default::default());
             rule.set_level(rule_severity.into());
         }