+++ crates/biome_configuration/src/linter/rules.rs
@@ -2714,8 +2714,8 @@ pub struct Nursery {
     pub use_array_literals: Option<RuleConfiguration<UseArrayLiterals>>,
     #[doc = "Enforce the use of new for all builtins, except String, Number, Boolean, Symbol and BigInt."]
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub use_consistent_builtin_instatiation:
-        Option<RuleConfiguration<UseConsistentBuiltinInstatiation>>,
+    pub use_consistent_builtin_instantiation:
+        Option<RuleConfiguration<UseConsistentBuiltinInstantiation>>,
     #[doc = "Require the default clause in switch statements."]
     #[serde(skip_serializing_if = "Option::is_none")]
     pub use_default_switch_clause: Option<RuleConfiguration<UseDefaultSwitchClause>>,
@@ -2770,7 +2770,7 @@ impl Nursery {
         "noUnknownUnit",
         "noUselessUndefinedInitialization",
         "useArrayLiterals",
-        "useConsistentBuiltinInstatiation",
+        "useConsistentBuiltinInstantiation",
         "useDefaultSwitchClause",
         "useExplicitLengthCheck",
         "useGenericFontNames",
@@ -2954,7 +2954,7 @@ impl Nursery {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[20]));
             }
         }
-        if let Some(rule) = self.use_consistent_builtin_instatiation.as_ref() {
+        if let Some(rule) = self.use_consistent_builtin_instantiation.as_ref() {
             if rule.is_enabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[21]));
             }
@@ -3093,7 +3093,7 @@ impl Nursery {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[20]));
             }
         }
-        if let Some(rule) = self.use_consistent_builtin_instatiation.as_ref() {
+        if let Some(rule) = self.use_consistent_builtin_instantiation.as_ref() {
             if rule.is_disabled() {
                 index_set.insert(RuleFilter::Rule(Self::GROUP_NAME, Self::GROUP_RULES[21]));
             }
@@ -3243,8 +3243,8 @@ impl Nursery {
                 .use_array_literals
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
-            "useConsistentBuiltinInstatiation" => self
-                .use_consistent_builtin_instatiation
+            "useConsistentBuiltinInstantiation" => self
+                .use_consistent_builtin_instantiation
                 .as_ref()
                 .map(|conf| (conf.level(), conf.get_options())),
             "useDefaultSwitchClause" => self