+++ crates/biome_graphql_parser/src/parser/mod.rs
@@ -15,6 +15,8 @@ use biome_parser::token_source::Trivia;
 use biome_parser::ParserContext;
 use definitions::DefinitionList;
 
+use self::value::{is_at_string, parse_string};
+
 pub(crate) struct GraphqlParser<'source> {
     context: ParserContext<GraphqlSyntaxKind>,
     source: GraphqlTokenSource<'source>,
@@ -30,6 +32,10 @@ impl<'source> GraphqlParser<'source> {
 
     pub fn lookahead(&mut self) -> GraphqlSyntaxKind {
         self.source.lookahead()
+    }
+
+    pub fn lookahead_at(&mut self, kind: GraphqlSyntaxKind) -> bool {
+        self.source.lookahead_at(kind)
     }
 
     pub fn finish(
@@ -90,6 +96,19 @@ fn parse_name(p: &mut GraphqlParser) -> ParsedSyntax {
     let m = p.start();
     p.bump(GRAPHQL_NAME);
     Present(m.complete(p, GRAPHQL_NAME))
+}
+
+#[inline]
+fn parse_description(p: &mut GraphqlParser) -> ParsedSyntax {
+    if !is_at_string(p) {
+        return Absent;
+    }
+
+    let m = p.start();
+    // already checked for in `is_at_string`
+    parse_string(p).ok();
+
+    Present(m.complete(p, GRAPHQL_DESCRIPTION))
 }
 
 #[inline]