+++ crates/biome_graphql_parser/src/parser/parse_error.rs
@@ -45,3 +45,14 @@ pub(crate) fn expected_argument(p: &GraphqlParser, range: TextRange) -> ParseDia
 pub(crate) fn expected_variable_definition(p: &GraphqlParser, range: TextRange) -> ParseDiagnostic {
     expected_node("variable definition", range, p)
 }
+
+pub(crate) fn expected_root_operation_type_definition(
+    p: &GraphqlParser,
+    range: TextRange,
+) -> ParseDiagnostic {
+    expected_node("root operation type definition", range, p)
+}
+
+pub(crate) fn expected_operation_type(p: &GraphqlParser, range: TextRange) -> ParseDiagnostic {
+    expected_any(&["query", "mutation", "subscription"], range, p)
+}